Updated test for relocated compond
