+++ qml/__init__.py
@@ -32,7 +32,6 @@ Provides
 """
 from __future__ import absolute_import
 
-from compound import Compound
 from . import models
 from . import aglaia
 from . import math
@@ -42,6 +41,7 @@ from . import fchl
 from . import representations
 from . import qmlearn
 from . import utils
+from .utils.compound import Compound
 
 __author__ = "Anders S. Christensen"
 __copyright__ = "Copyright 2016"
