+++ test/test_arad.py
@@ -72,7 +72,7 @@ def test_arad():
     for xyz_file in sorted(data.keys())[:10]:
 
         # Initialize the qml.data.Compound() objects
-        mol = qml.data.Compound(xyz=test_dir + "/qm7/" + xyz_file)
+        mol = qml.Compound(xyz=test_dir + "/qm7/" + xyz_file)
 
         # Associate a property (heat of formation) with the object
         mol.properties = data[xyz_file]
