+++ test/test_compound.py
@@ -24,7 +24,7 @@ from __future__ import print_function
 
 import os
 
-from qml.data import Compound
+from qml import Compound
 import numpy as np
 
 def compare_lists(a, b):
